+++ asn1/api/lc_x509_parser.h
@@ -185,6 +185,27 @@ int lc_x509_cert_decode(struct lc_x509_certificate *cert, const uint8_t *data,
 	(name)->data_struct_size = LC_X509_KEYS_DILITHIUM_ED25519_SIZE;        \
 	_Pragma("GCC diagnostic pop")
 #endif
+
+#ifdef LC_DILITHIUM_ED448_SIG
+#define LC_X509_KEYS_DILITHIUM_ED448_SIZE                                    \
+	(sizeof(struct lc_dilithium_ed448_pk) +                              \
+	 sizeof(struct lc_dilithium_ed448_sk) +                              \
+	 sizeof(struct lc_x509_key_data))
+#define LC_X509_KEYS_DILITHIUM_ED448_SET(name)                               \
+	_Pragma("GCC diagnostic push")                                         \
+		_Pragma("GCC diagnostic ignored \"-Wcast-align\"")(name)       \
+			->pk.dilithium_ed448_pk =                            \
+		(struct lc_dilithium_ed448_pk                                \
+			 *)((uint8_t *)(name) +                                \
+			    sizeof(struct lc_x509_key_data));                  \
+	(name)->sk.dilithium_ed448_sk =                                      \
+		(struct lc_dilithium_ed448_sk                                \
+			 *)((uint8_t *)(name) +                                \
+			    sizeof(struct lc_x509_key_data) +                  \
+			    sizeof(struct lc_dilithium_ed448_pk));           \
+	(name)->data_struct_size = LC_X509_KEYS_DILITHIUM_ED448_SIZE;        \
+	_Pragma("GCC diagnostic pop")
+#endif
 /// \endcond
 
 /**
@@ -208,12 +229,34 @@ int lc_x509_cert_decode(struct lc_x509_certificate *cert, const uint8_t *data,
 
 /**
  * @ingroup X509
+ * @brief Allocate memory for struct lc_x509_keys_data holding Dilithium-ED448
+ *	  keys on stack
+ *
+ * @param [in] name Name of stack variable
+ */
+#define LC_X509_KEYS_DILITHIUM_ED448_ON_STACK(name)                          \
+	_Pragma("GCC diagnostic push") _Pragma(                                \
+		"GCC diagnostic ignored \"-Wdeclaration-after-statement\"")    \
+		_Pragma("GCC diagnostic ignored \"-Wcast-align\"")             \
+			LC_ALIGNED_BUFFER(name##_ctx_buf,                      \
+					  LC_X509_KEYS_DILITHIUM_ED448_SIZE, \
+					  LC_HASH_COMMON_ALIGNMENT);           \
+	struct lc_x509_key_data *name =                                        \
+		(struct lc_x509_key_data *)name##_ctx_buf;                     \
+	LC_X509_KEYS_DILITHIUM_ED448_SET(name);                              \
+	_Pragma("GCC diagnostic pop")
+
+/**
+ * @ingroup X509
  * @brief Allocate memory for struct lc_x509_keys_data holding any kind of key
  *	  type on stack
  *
  * @param [in] name Name of stack variable
  */
-#ifdef LC_DILITHIUM_ED25519_SIG
+#ifdef LC_DILITHIUM_ED448_SIG
+#define LC_X509_KEYS_ON_STACK(name)                                            \
+	LC_X509_KEYS_DILITHIUM_ED448_ON_STACK(name)
+#elif defined LC_DILITHIUM_ED25519_SIG
 #define LC_X509_KEYS_ON_STACK(name)                                            \
 	LC_X509_KEYS_DILITHIUM_ED25519_ON_STACK(name)
 #elif defined(LC_SPHNCS_ENABLED)
@@ -247,6 +290,17 @@ static inline void lc_x509_keys_zero(struct lc_x509_key_data *keys)
  * @return 0 on success or < 0 on error
  */
 int lc_x509_keys_dilithium_ed25519_alloc(struct lc_x509_key_data **keys);
+
+/**
+ * @ingroup X509
+ * @brief Allocate memory for struct lc_x509_keys_data holding Dilithium-ED448
+ *	  keys on heap
+ *
+ * @param [in] keys Variable to allocate
+ *
+ * @return 0 on success or < 0 on error
+ */
+int lc_x509_keys_dilithium_ed448_alloc(struct lc_x509_key_data **keys);
 
 /**
  * @ingroup X509
@@ -376,6 +430,20 @@ int lc_x509_signature_verify(const uint8_t *sig_data, size_t siglen,
  */
 int lc_x509_cert_load_pk_dilithium_ed25519(
 	struct lc_dilithium_ed25519_pk *dilithium_ed25519_pk,
+	const uint8_t *pk_ptr, size_t pk_len);
+
+/**
+ * @ingroup X509
+ * @brief Parse a Composite ML-DSA ASN.1 structure into a public key structure
+ *
+ * @param [out] dilithium_ed448_pk Public key to be filled
+ * @param [in] pk_ptr Pointer to ASN.1 structure
+ * @param [out] pk_len Size of the public key ASN.1 structure
+ *
+ * @return 0 on success or < 0 on error
+ */
+int lc_x509_cert_load_pk_dilithium_ed448(
+	struct lc_dilithium_ed448_pk *dilithium_ed448_pk,
 	const uint8_t *pk_ptr, size_t pk_len);
 
 /**